pyproject.toml
[project]
name = "absa-sentiment-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv-python==0.0.1",
    "google-generativeai==0.8.5",
    "google-play-scraper==1.2.7",
    "ipykernel==6.30.1",
    "pandas==2.3.2",
    "python-dotenv==1.1.1",
    "tqdm==4.67.1",
]
